STABLE12-afs-attrs-cast-date-to-int-to-avoid-negative-dates-20020516
cheap hack to avoid negative date problem which occasionally plagues solaris


(cherry picked from commit c28c22ceadfb1db8151ae74dacee2a49c42624db)

+++ src/afs/VNOPS/afs_vnop_attrs.c
@@ -93,7 +93,7 @@ afs_CopyOutAttrs(avc, attrs)
     attrs->va_nlink = avc->m.LinkCount;
     attrs->va_size = avc->m.Length;
     attrs->va_atime.tv_sec = attrs->va_mtime.tv_sec = attrs->va_ctime.tv_sec =
-	avc->m.Date;
+	(int)avc->m.Date;
     /* set microseconds to be dataversion # so that we approximate NFS-style
      * use of mtime as a dataversion #.  We take it mod 512K because
      * microseconds *must* be less than a million, and 512K is the biggest
